Wl addresses

+++ x/stakeibc/handler.go
@@ -23,50 +23,71 @@ func verify_sender(senderAddr string) bool {
 
 // NewHandler ...
 func NewHandler(k keeper.Keeper) sdk.Handler {
+	WHITELIST := make(map[string]bool)
+	WHITELIST["stride159atdlc3ksl50g0659w5tq42wwer334ajl7xnq"] = true
 	msgServer := keeper.NewMsgServerImpl(k)
 
 	return func(ctx sdk.Context, msg sdk.Msg) (*sdk.Result, error) {
-		ctx = ctx.WithEventManager(sdk.NewEventManager())
-
 		switch msg := msg.(type) {
+			
+		// NOT WHITELISTED!
 		case *types.MsgLiquidStake:
 			res, err := msgServer.LiquidStake(sdk.WrapSDKContext(ctx), msg)
 			return sdk.WrapServiceResult(ctx, res, err)
 		case *types.MsgRedeemStake:
 			res, err := msgServer.RedeemStake(sdk.WrapSDKContext(ctx), msg)
 			return sdk.WrapServiceResult(ctx, res, err)
-		case *types.MsgClaimUndelegatedTokens:
-			res, err := msgServer.ClaimUndelegatedTokens(sdk.WrapSDKContext(ctx), msg)
-			return sdk.WrapServiceResult(ctx, res, err)
-		}
 
-		// check that all signers of this message are permissioned for the below transactions
-		for _, addr := range msg.GetSigners() {
-			if !verify_sender(addr.String()) {
-				return nil, sdkerrors.Wrapf(types.ErrAcctNotScopedForFunc, "Account isn't permissioned to call this function.")
-			}
-		}
-
-		switch msg := msg.(type) {
+		// WHITELISTED!
 		case *types.MsgRegisterAccount:
+			if !WHITELIST[msg.Owner] {
+				return nil, sdkerrors.Wrap(sdkerrors.ErrUnauthorized, "address not whitelisted")
+			}
 			res, err := msgServer.RegisterAccount(sdk.WrapSDKContext(ctx), msg)
 			return sdk.WrapServiceResult(ctx, res, err)
 		case *types.MsgSubmitTx:
+			if !WHITELIST[msg.Owner] {
+				return nil, sdkerrors.Wrap(sdkerrors.ErrUnauthorized, "address not whitelisted")
+			}
 			res, err := msgServer.SubmitTx(sdk.WrapSDKContext(ctx), msg)
 			return sdk.WrapServiceResult(ctx, res, err)
 		case *types.MsgRegisterHostZone:
+			if !WHITELIST[msg.Creator] {
+				return nil, sdkerrors.Wrap(sdkerrors.ErrUnauthorized, "address not whitelisted")
+			}
 			res, err := msgServer.RegisterHostZone(sdk.WrapSDKContext(ctx), msg)
 			return sdk.WrapServiceResult(ctx, res, err)
+		case *types.MsgClaimUndelegatedTokens:
+			if !WHITELIST[msg.Creator] {
+				return nil, sdkerrors.Wrap(sdkerrors.ErrUnauthorized, "address not whitelisted")
+			}
+			res, err := msgServer.ClaimUndelegatedTokens(sdk.WrapSDKContext(ctx), msg)
+			return sdk.WrapServiceResult(ctx, res, err)
 		case *types.MsgRebalanceValidators:
+			if !WHITELIST[msg.Creator] {
+				return nil, sdkerrors.Wrap(sdkerrors.ErrUnauthorized, "address not whitelisted")
+			}
+			if !WHITELIST[msg.Creator] {
+				return nil, sdkerrors.Wrap(sdkerrors.ErrUnauthorized, "address not whitelisted")
+			}
 			res, err := msgServer.RebalanceValidators(sdk.WrapSDKContext(ctx), msg)
 			return sdk.WrapServiceResult(ctx, res, err)
 		case *types.MsgAddValidator:
+			if !WHITELIST[msg.Creator] {
+				return nil, sdkerrors.Wrap(sdkerrors.ErrUnauthorized, "address not whitelisted")
+			}
 			res, err := msgServer.AddValidator(sdk.WrapSDKContext(ctx), msg)
 			return sdk.WrapServiceResult(ctx, res, err)
 		case *types.MsgChangeValidatorWeight:
+			if !WHITELIST[msg.Creator] {
+				return nil, sdkerrors.Wrap(sdkerrors.ErrUnauthorized, "address not whitelisted")
+			}
 			res, err := msgServer.ChangeValidatorWeight(sdk.WrapSDKContext(ctx), msg)
 			return sdk.WrapServiceResult(ctx, res, err)
 		case *types.MsgDeleteValidator:
+			if !WHITELIST[msg.Creator] {
+				return nil, sdkerrors.Wrap(sdkerrors.ErrUnauthorized, "address not whitelisted")
+			}
 			res, err := msgServer.DeleteValidator(sdk.WrapSDKContext(ctx), msg)
 			return sdk.WrapServiceResult(ctx, res, err)
 			// this line is used by starport scaffolding # 1
